+++ plugins/blog/plugin-comment/src/client/components/WalineComment.ts
@@ -24,6 +24,8 @@ import '@waline/client/waline.css'
 import '../styles/waline.scss'
 
 declare const WALINE_META: boolean
+// FIXME: https://github.com/oxc-project/tsgolint/issues/861
+// oxlint-disable-next-line typescript/no-unnecessary-type-arguments
 declare const WALINE_LOCALES: ExactLocaleConfig<WalineLocaleData>
 
 const walineLocales = WALINE_LOCALES