+++ plugins/markdown/plugin-prismjs/src/node/parser/highlightLines.ts
@@ -21,7 +21,7 @@ export const getHighlightLinesRange = (
   info: string,
 ): HighlightLinesRange[] | null => {
   // try to match highlight-lines mark
-  const match = info.match(/{([\d,-]+)}/)
+  const match = /{([\d,-]+)}/.exec(info)
 
   // no highlight-lines mark, return `null`
   if (match == null) return null