+++ plugins/search/plugin-slimsearch/tsdown.config.ts
@@ -6,14 +6,14 @@ export default [
     'client/config',
     'client/index',
     'client/shims.d',
-    'client/worker',
+    'worker/dev',
   ]),
-  tsdownConfig('worker/index', {
+  tsdownConfig('worker/build', {
     dts: false,
     define: {
       '__VUEPRESS_SSR__': 'false',
       'process.env.NODE_ENV': JSON.stringify('production'),
     },
-    alwaysBundle: [/^@vuepress\//, 'slimsearch'],
+    alwaysBundle: [/^@vuepress\//, 'slimsearch', /^vuepress\//],
   }),
 ]