+++ themes/theme-default/src/client/utils/resolveRepoType.ts
@@ -3,9 +3,12 @@ import { isLinkHttp } from 'vuepress/shared'
 export type RepoType = 'Bitbucket' | 'Gitee' | 'GitHub' | 'GitLab' | null
 
 export const resolveRepoType = (repo: string): RepoType => {
-  if (!isLinkHttp(repo) || repo.includes('github.com')) return 'GitHub'
-  if (repo.includes('bitbucket.org')) return 'Bitbucket'
-  if (repo.includes('gitlab.com')) return 'GitLab'
-  if (repo.includes('gitee.com')) return 'Gitee'
+  if (!isLinkHttp(repo) || /^(https?:)?\/\/github\.com\//.test(repo))
+    return 'GitHub'
+
+  if (/^(https?:)?\/\/bitbucket\.org\//.test(repo)) return 'Bitbucket'
+  if (/^(https?:)?\/\/gitlab\.com\//.test(repo)) return 'GitLab'
+  if (/^(https?:)?\/\/gitee\.com\//.test(repo)) return 'Gitee'
+
   return null
 }